refactor: renamed LoginScreen.dart to login_screen.dart
